Moved RegisterInputValidatorTest to correct unit testing package
